PYIC-3472: Refactor vchelper to gpg45-evaluator library
